unified copyright statements, cleaned up temporal methods
