+++ LICENSE
@@ -1,8 +1,7 @@
 MIT License
 
-Copyright (c) 2019- Michał Czuba, Piotr Bródka
-Copyright (c) 2023 Damian Dąbrowski, Damian Serwata, Radosław Michalski
-Copyright (c) 2023- Mateusz Nurek, Yu-Xuan Qi, Mingshan Jia, Katarzyna Musial
+Copyright (c) 2025 Piotr Bródka, Michał Czuba, Damian Dąbrowski, Mingshan Jia
+Radosław Michalski, Katarzyna Musial, Mateusz Nurek, Yu-Xuan Qi, Damian Serwata
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
