+++ network_diffusion/seeding/closeness_selector.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2023 by Damian Dąbrowski, Michał Czuba.
+# Copyright (c) 2025 by Damian Dąbrowski, Michał Czuba.
 #
 # This file is a part of Network Diffusion.
 #
