+++ network_diffusion/seeding/kppshell_selector.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2024 by Michał Czuba, Piotr Bródka.
+# Copyright (c) 2025 by Michał Czuba, Piotr Bródka.
 #
 # This file is a part of Network Diffusion.
 #
